==> Part 2 Clean Data.ipynb <==
import pandas as pd
from bs4 import BeautifulSoup
import re

train_raw = pd.read_csv("train_raw.csv")
test_raw = pd.read_csv("test_raw.csv")

def clean_html(review) : 
    text = BeautifulSoup(review, "html.parser").get_text() # Remove HTML tags
    text = re.sub(r"[^a-zA-Z0-9]", " ", text.lower()) # Convert to lower case
    return text

train_raw['clean_text'] = train_raw["text"].apply(clean_html)
test_raw['clean_text'] = test_raw["text"].apply(clean_html)

train_raw.head()

test_raw.head()

del train_raw['text']
del test_raw['text']

train_raw.to_csv("train_clean.csv",index=False)
test_raw.to_csv("test_clean.csv",index=False)
