==> Part 3 Separate Clean text to row.ipynb <==
import pandas as pd
import nltk
import pickle


nltk.download("stopwords")
from nltk.corpus import stopwords
from nltk.stem.porter import *

stemmer = PorterStemmer()
stopwords_english = stopwords.words("english")

train_clean = pd.read_csv("train_clean.csv")
test_clean = pd.read_csv("test_clean.csv")

train_clean.head()

def review_to_words(text):
    words = text.split() # Split string into words
    words = [w for w in words if w not in stopwords_english] # Remove stopwords
    words = [PorterStemmer().stem(w) for w in words] # stem
    return words

def review_to_words2(text):
    words = text.split() # Split string into words
    print("words1 : ", words)
    words = [w for w in words if w not in stopwords_english] # Remove stopwords
    print("words2 : ", words)
    
    final_words = []
    for w in words :
        result = PorterStemmer().stem(w)
        print("result :", result)
        final_words.append(result)
    
    print("final_words : ", final_words)
    return final_words

sample_train = train_clean["clean_text"][0]

sample_train = "This is not good movie"

result_train = review_to_words2(sample_train)
result_train

train_clean['final_text'] = train_clean["clean_text"].apply(review_to_words)
test_clean['final_text'] = test_clean["clean_text"].apply(review_to_words)

train_clean.head()

del train_clean['clean_text']
del test_clean['clean_text']

train_clean.to_csv("train_words.csv",index=False)
test_clean.to_csv("test_words.csv",index=False)

cache_data = {
    "train_words":train_clean, 
    "test_words":test_clean
}

with open('data_words.pickle', "wb") as f:
                pickle.dump(cache_data, f)
