==> Part 4 Transform Feature.ipynb <==
import pandas as pd
import pickle

import numpy as np
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.externals import joblib

with open('data_words.pickle', 'rb') as handle:
    data_words = pickle.load(handle)

train_words = data_words["train_words"]
test_words = data_words["test_words"]

train_words.head()

# SAMPLE CORPUS

corpus = [
'This is the first document.',
'This document is the second document.',
'And this is the third one.',
'Is this the first document?',
]


sample_vectorizer = CountVectorizer()

X = sample_vectorizer.fit_transform(corpus)

print(sample_vectorizer.get_feature_names())
print(X.toarray())

sample_vectorizer2 = CountVectorizer(analyzer='word', ngram_range=(2, 2))
X2 = sample_vectorizer2.fit_transform(corpus)
print(sample_vectorizer2.get_feature_names())
print(X2.toarray())

# SAMPLE CORPUS DONE

vocabulary_size = 500

vectorizer = CountVectorizer(
    max_features=vocabulary_size,
    preprocessor=lambda x: x, 
    tokenizer=lambda x: x
)  # 

features_train = vectorizer.fit_transform(train_words["final_text"]).toarray()
features_test = vectorizer.transform(test_words["final_text"]).toarray()

features_train

train_words.shape

train_words["features"] = list(features_train)
test_words["features"] = list(features_test)

train_words.head()

test_words.head()

cache_data = {
    "train_transform":train_words, 
    "test_transform":test_words,
    "vocabulary_size":vocabulary_size
}

with open('data_features.pickle', "wb") as f:
                pickle.dump(cache_data, f)
