==> Part 5 Logistic Regresion.ipynb <==
import numpy as np
import pandas as pd
import pickle

from sklearn.linear_model import LinearRegression,Ridge,Lasso
from sklearn.svm import SVR
from xgboost import XGBRegressor
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier

from sklearn.tree import DecisionTreeClassifier
from sklearn.neighbors import KNeighborsClassifier
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
from sklearn.naive_bayes import GaussianNB


from sklearn.metrics import accuracy_score
from sklearn.metrics import precision_score
from sklearn.metrics import recall_score
from sklearn.metrics import f1_score
from sklearn.metrics import roc_curve, auc
from sklearn.metrics import roc_auc_score

with open('data_features.pickle', 'rb') as handle:
    data_features = pickle.load(handle)

train_data = data_features["train_transform"]
test_data = data_features["test_transform"]

train_data.head()

test_data.head()

train_raw = pd.DataFrame(list(train_data["features"])) 
train_raw["target"] = train_data["target"]

test_raw = pd.DataFrame(list(test_data["features"])) 
test_raw["target"] = test_data["target"]


print("train_raw : ", train_raw.shape)
train_raw.head()

print("test_raw : ", test_raw.shape)
test_raw.head()

X_train = train_raw.loc[:, train_raw.columns != 'target']
Y_train = train_raw.loc[:, train_raw.columns == 'target']

X_test = test_raw.loc[:, test_raw.columns != 'target']
Y_test = test_raw.loc[:, test_raw.columns == 'target']

X_train.head()

from sklearn.ensemble import RandomForestClassifier


selected_model = {
    "LinearRegression" : LinearRegression(),
#     "Ridge" : Ridge(),
#     "Lasso" : Lasso(),
    "XGBRegressor" : XGBRegressor(),
#     "SVM" : SVR(),
    "LogisticRegression" : LogisticRegression(),
    "DecisionTreeClassifier" : DecisionTreeClassifier(),
    "RandomForestClassifier":RandomForestClassifier()
#     "KNeighborsClassifier" : KNeighborsClassifier(),
#     "LinearDiscriminantAnalysis" : LinearDiscriminantAnalysis(),
#     "GaussianNB" : GaussianNB()
    
    
}

def model_evaluate() :
    
    for model_name,model_value in selected_model.items() :
        print("running model : ", model_name)
        model = model_value
        model.fit(X_train,Y_train)
        target_predicted = model.predict(X_test)
        # print(model_name, " target_predicted : ", target_predicted)
        acc = accuracy_score(Y_test["target"], target_predicted.round())

        
        
        print(model_name," : ", acc)

model_evaluate()

model = LogisticRegression(n_jobs=5,max_iter=200)
model

model.fit(X_train,Y_train)
Y_test_predicted = model.predict(X_test)

Y_test["target_predicted"] = Y_test_predicted.round()

Y_test.head()

acc = accuracy_score(Y_test["target"], Y_test["target_predicted"].round())
precision = precision_score(Y_test["target"], Y_test["target_predicted"],average="weighted")
recall = recall_score(Y_test["target"], Y_test["target_predicted"],average="weighted")
f1 = f1_score(Y_test["target"], Y_test["target_predicted"],average="weighted")
roc = roc_auc_score(Y_test["target"], Y_test["target_predicted"].round())

print("acc : ", acc)
print("precision : ", precision)
print("recall : ", recall)
print("f1 : ", f1)
print("roc : ", roc)
